refactor files for easier maintance
